removed bar folder and reworked pathing for icons
