+++ src/App.tsx
@@ -1,11 +1,13 @@
 import Header from "./ui/Header";
 import Tags from "./ui/Tags";
+import Sidebar from "./ui/Header/Sidebar";
 
 const App = () => {
   return (
     <div className="App">
       <Header />
       <Tags />
+      <Sidebar />
     </div>
   );
 };
